+++ jin/src/main/java/springbook/user/dao/DaoFactory.java
@@ -1,0 +1,9 @@
+package springbook.user.dao;
+
+public class DaoFactory {
+    public UserDao userDao() {
+        ConnectionMaker connectionMaker = new DConnectionMaker();
+        UserDao userDao = new UserDao(connectionMaker);
+        return userDao;
+    }
+}
